# gen/anonde/v1/anonde.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "anonde/v1/anonde.proto",
    "description": "Anonde service API.\n\nv1 is not yet released; breaking changes are allowed within this version\nuntil the first tagged release. Once tagged, follow the usual proto\nrules (no field renumbering, no required→optional flips, etc.).\n\nTwo transports share these messages:\n\n  - gRPC / Connect (RPC-style URLs):\n      POST /anonde.v1.Service/\u003cMethod\u003e\n\n  - REST gateway (grpc-gateway, path-based URLs):\n      see the (google.api.http) option on each rpc below.\n\nTenant/anonymization identifiers live in the URL path / body on the\nREST side and inside the request message on the RPC side. The\ngateway populates the message fields from the path before\ndispatching, so the service layer only ever sees one shape.",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/anonymizations/{id}/reveal-pdf": {
      "get": {
        "summary": "RevealPDF returns the original bytes the server stashed at\nAnonymizePDF time. 404 when the record is missing, expired, or was\nnot created via the PDF endpoint (i.e. no OriginalBytes on the\nstore record). Same tenant binding as AnonymizePDF.",
        "description": "REST: GET /v1/anonymizations/{id}/reveal-pdf",
        "operationId": "Service_RevealPDF",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RevealPDFResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Service"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1RevealPDFResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "originalPdf": {
          "type": "string",
          "format": "byte",
          "description": "Original (pre-anonymization) PDF bytes the server stashed at\nAnonymizePDF time."
        }
      }
    },
...
  }
}